requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
groq==0.9.0
supabase==2.4.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
apscheduler==3.10.4
httpx==0.27.0
PyMuPDF==1.24.3
reportlab==4.1.0
python-dotenv==1.0.1
